+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email "andrea.campi@zephirworks.com"
 license          "Apache 2.0"
 description      "Deploys and configures Java-based applications"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "1.0.0"
+version          "1.0.2"
 
 %w{ application java tomcat }.each do |cb|
   depends cb