+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v1.0.2:
+
+* Support deploying local files
+
+## v1.0.1:
+
+* [COOK-1680] - Add deploy action (alias for create) to java_remote_file
+
 ## v1.0.0:
 
 * [COOK-1243] - Initial release - relates to COOK-634.